requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
requests==2.32.3
pydantic==2.9.2
pypdf==5.1.0
python-dotenv==1.0.1
